@@ -32,56 +32,47 @@
     "component"
   ],
   "scripts": {
-    "lint": "tslint --force \"src/**/*.ts\"",
-    "pretest": "npm run lint",
-    "test": "karma start",
-    "test-watch": "karma start --no-single-run --auto-watch",
-    "build": "ng-packagr -p ng-package.json"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
   "devDependencies": {
-    "@angular/common": "^5.2.1",
-    "@angular/compiler": "^5.2.1",
-    "@angular/compiler-cli": "^5.2.1",
-    "@angular/core": "^5.2.1",
-    "@angular/http": "^5.2.1",
-    "@angular/platform-browser": "^5.2.1",
-    "@angular/platform-browser-dynamic": "^5.2.1",
-    "@angular/router": "^5.2.1",
-    "@ngtools/webpack": "^1.2.9",
-    "@types/jasmine": "^2.5.42",
-    "@types/node": "^9.3.0",
-    "angular-router-loader": "^0.8.2",
-    "angular2-template-loader": "^0.6.2",
-    "codelyzer": "^4.1.0",
-    "css-loader": "^0.28.0",
-    "file-loader": "^1.1.6",
-    "html-loader": "^0.5.1",
-    "istanbul-instrumenter-loader": "^3.0.0",
-    "jasmine-core": "^2.5.2",
-    "json-loader": "^0.5.3",
-    "karma": "^2.0.0",
-    "karma-chrome-launcher": "2.2.0",
-    "karma-coverage-istanbul-reporter": "^1.3.3",
-    "karma-jasmine": "^1.1.1",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-spec-reporter": "0.0.32",
-    "karma-webpack": "^2.0.9",
-    "less": "^2.7.1",
-    "less-loader": "^4.0.3",
-    "ng-packagr": "^2.2.0",
-    "raw-loader": "0.5.1",
-    "reflect-metadata": "^0.1.9",
-    "rimraf": "^2.6.1",
-    "rxjs": "^5.5.7",
-    "source-map-loader": "0.2.3",
-    "to-string-loader": "^1.1.5",
-    "ts-loader": "^3.2.0",
-    "tsickle": "^0.26.0",
-    "tslint": "^5.0.0",
-    "tslint-loader": "^3.5.1",
-    "typescript": "^2.4.0",
-    "url-loader": "^1.0.1",
-    "webpack": "^3.0.0",
-    "zone.js": "^0.8.5"
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
+    "core-js": "^2.5.4",
+    "rxjs": "~6.2.0",
+    "zone.js": "~0.8.26",
+    "@angular-devkit/build-angular": "~0.8.0",
+    "@angular-devkit/build-ng-packagr": "~0.8.0",
+    "@angular/cli": "~6.2.3",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "codelyzer": "~4.3.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^4.1.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tsickle": ">=0.29.0",
+    "tslib": "^1.9.0",
+    "tslint": "~5.11.0",
+    "typescript": "~2.9.2"
   }
 }
