@@ -1,5 +1,5 @@
-export * from './components';
-export * from './directives';
-export * from './navigation';
-export * from './util';
-export {ArchwizardModule} from './archwizard.module';
+export * from './lib/components';
+export * from './lib/directives';
+export * from './lib/navigation';
+export * from './lib/util';
+export {ArchwizardModule} from './lib/archwizard.module';
